requirements.txt
svgpathtools
pyclipper
scipy
torchvision
editdistance
numpy
opencv-python
pillow
PyYAML
torch
